+++ src/itemsList/ItemsList.style.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const StyledItemsList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  list-style-type: none;
+`